packaging: Added postgres listening external network
This patch adds the interaction for the user allowing
to setup the local postgres to listen on external
connections.

The patch will check if it is able to create database, and
will only offer to configure postgres in local installation. It
will also create a read-only role with user-provided
credentials and open the required firewall ports if necessary.

It also creates a self-signed certificate if user choses to
use secure connections.

Change-Id: Ibdd133a392fb3dba6088a9fb12ce4ca828f24073
Bug-Url: https://bugzilla.redhat.com/737598
Signed-off-by: Alex Lourie <alourie@redhat.com>