packaging: services: remove pid when service start fails
the service is running under non privileged user, so it cannot remove
the pid file only clear it.

if service is dead at validation stage its status is printed as:

  dead but pid file exists

to remedy that, if service fails to start we remove the pid if service
start fails and pid file is zero length.

Change-Id: I11a29e58e0fd673eb1df79983b0c09799f9ef0f3
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>