packaging: setup: postgres95: Fixes
- Upgrade in main transaction

Before this patch, postgres upgrade was done in its own transaction.
So rollback could only happen if there was a failure in the upgrade
process.

Move this to otopi's MAIN_TRANSACTION, which is where most changes are
done - so that if we fail elsewhere and rollback, we also rollback
postgresql.

- Do not backup/restore on upgrade

Not needed, because we can revert to the old postgresql for rollback.

Change-Id: Ieeace3bb4c784ab8090b43dafcd8150b0f461ce4
Bug-Url: https://bugzilla.redhat.com/1546487
Signed-off-by: Yedidyah Bar David <didi@redhat.com>