Add grafana integration support
Initial commit to add grafana integration.

Only tested with engine+dwh+grafana+DBs on same machine, clean setup and
upgrade.

Some other configurations are in principle supported in the code, but
not tested. Others are not supported, and some of these cases are marked
with 'TODO' comments in the code.

Bug-Url: https://bugzilla.redhat.com/1814643
Change-Id: Iba6ed351f3c8f72408e281b5e69ab45cbc707381
Signed-off-by: Yedidyah Bar David <didi@redhat.com>