packaging: setup: Fix grafana pki conf after dwh
We have a single env key, REMOTE_ENGINE_CONFIGURED, for marking whether
a remote engine was already configured. Originally it was used only for
DWH. Later we used it also for grafana. But this flow does not work:

1. Configure engine on machine A
2. Configure DWH on machine B
3. Configure grafana on machine B

Because REMOTE_ENGINE_CONFIGURED was already set, so we do not ask the
user for relevant information, and fail.

Fix this by clearing REMOTE_ENGINE_CONFIGURED if user asked to configure
grafana but pki is missing.

Change-Id: Ic7c4cab41f27b02bddf7c67b0985e743833cfdd7
Bug-Url: https://bugzilla.redhat.com/1919984
Signed-off-by: Yedidyah Bar David <didi@redhat.com>