packaging: dbscripts: Improve logging
This is a combination of two engine patches [1][2], slightly fixed so
that they apply to dwh.

[1] engine commit hash cd728dae, change id Id6fc86617396e23c16ff8be9fa9

dbscripts: Output names of skipped files

Only upgrade scripts with version higher than "current" (last script
ran) are ran. Output those we skip.

[2] engine commit hash 6d88035, change id I29e558b6b14564a223b5baca763f

packaging: dbscripts: Improve logging

Output the various "titles" to the log file, instead of to stdout.

We call psql with --log-file=DBFUNC_LOGFILE, so it logs to there
immediately. But the normal output (to stdout) goes there only by
engine-setup (otopi), after schema.sh finishes, so it's impossible to
match the title with the psql output. So:

Replace all relevant calls to 'echo' with calls to a new function
dbfunc_output that appends to DBFUNC_LOGFILE if set, otherwise to
stdout.

Replace most calls to dbfunc_psql and dbfunc_psql_die with calls to new
functions dbfunc_psql_v and dbfunc_psql_die_v that log/output that they
were called.

Instead of pipes like 'psql|psql', save the output to a variable
and call on it.

Change-Id: Icd0e983ab5441eb7fc41a540882e2e69f37d101c
Signed-off-by: Yedidyah Bar David <didi@redhat.com>