packaging: setup: Usually default to configure grafana
We used to configure grafana by default.

This broke upgrade/restore of hosted-engine from 4.3 if the
backed up engine was set up with a remote DWH database[1].

To fix this, we kept the default in DWH to True, but patched
the appliance to override this to False, unconditionally.
This means that hosted-engine setups never configured grafana
out-of-the-box, and always required configuring it manually
later if needed, which is sub-optimal for most users.

We want to configure grafana by default in more cases. So:
Default to True (Configure grafana), unless all of the following
are true:
1. The DWH database is not new/empty
2. The DWH database host is not 'localhost'
3. OTOPI's "accept defaults" flag is not used ('--accept-defaults'
in the engine-setup command line, or
'DIALOG/autoAcceptDefault=bool:True' in the answer file)

I hope this is enough to keep [1] fixed but configure grafana in most
other cases.
[1] https://bugzilla.redhat.com/show_bug.cgi?id=1866780

Change-Id: I059f28114eb8189afc4facb01bc63bce2c83bcda
Bug-Url: https://bugzilla.redhat.com/1980315
Signed-off-by: Yedidyah Bar David <didi@redhat.com>