Fix vm interface queries
Update the joins to be the same as in the db.
This will maintain consistency between the db and Grafana,
And we will lower the risk that the query will fail
due to columns that can be null.

Change-Id: Ie99a800c7e7b25bbde1887698c57bdfcf94727e8
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=2014883
Signed-off-by: Aviv Litman <alitman@redhat.com>
(cherry picked from commit 359a198aa85cf6a0c240055e0012813c6ae3389c)