packaging: setup: grafana: Fix setting default privs
On new setups, the grafana db user was created only after db schema was
updated, but a recent patch tried to set default privileges to this user
before that, and this failed.

Originally, I created the user only after db schema was created because
the function to create the user also granted it access to the db. It
also tried to set default privs so that it has access to future tables,
but this never worked, as we discovered only on the move to 4.5, where
we actually added tables. Now that we fixed the default privileges
issue, it should be ok to create the user before the db, then set
default privs, then create the schema.

Bug-Url: https://bugzilla.redhat.com/2026358
Change-Id: I7563fcede75052b9122ad47cc9383e46a29cf845
Signed-off-by: Yedidyah Bar David <didi@redhat.com>