packaging: setup: Fix configuring grafana with keycloak
With this flow:
1. Setup an engine with keycloak, no dwh/grafana
2. engine-setup --reconfigure-optional-components, enable dwh+grafana

keycloak is already configured, but we do want to configure grafana to
use keycloak. So remove the condition that we only do this if keycloak
is not configured yet.

Also support configuring grafana on a separate machine when keycloak on
the engine machine is configured.

Requires [1]. spec updated to require 4.5.4 engine-setup to get it.

[1] https://github.com/oVirt/ovirt-engine/pull/669

Change-Id: I019335c60f686d139da0e9368147e035db75ba37
Signed-off-by: Yedidyah Bar David <didi@redhat.com>