etltermination: wait for main thread to end
We should wait until the main thread ends.
Otherwise we stop too early,
and not everything is correctly handled.

For example the DwhCurrentlyRunning value in the database is not changed
because of this. This makes engine-setup fail, as it thinks DWH is still
running.

Therefor we add a shutdown hook which waits until the main thread
finished everything (the main thread also waits for the other threads to
finish first), and then we can safely end.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>