service: add graceful shutdown
Attempt to graceful shutdown when signal is received.
i.e. instead of sending a signal to jvm, set DisconnectDwh to 1
in the engine database. This will make dwh cleanly finish what it
is doing and shut down. At startup we set the DisconnectDwh back
to 0, otherwise dwh will immediately be stopped again.

Signed-off-by: Peter Boden <peter.boden@team.blue>