service: let shutdownhook wait for stop
Relying on `DisconnectDwh` and receiving a signal at the
same time doesn't work. Systemd relays the signal to the
whole control-group, not just the python wrapper. This creates
a race condition in which the cleanup doesn't get executed most of
the time. Because of this the `DwhCurrentlyRunning` flag stays
set to `1`, which makes the ovirt-engine-setup fail.

New aproach:
- shutdownhook set terminate to true
- this triggers the talend code to stop
- shutdownhook then waits for stopped boolean to be true
- stopped is set to true right before calling `System.exit(exitCode)`

Signed-off-by: Peter Boden <peter.boden@team.blue>